Fix for hbase stat calculation.

+++ contrib/storage-hbase/src/main/java/org/apache/drill/exec/store/hbase/TableStatsCalculator.java
@@ -155,7 +155,7 @@ public class TableStatsCalculator {
    */
   public long getRegionSizeInBytes(byte[] regionId) {
     if (sizeMap == null) {
-      return avgRowSizeInBytes*1024*1024; // 1 million rows
+      return (long)avgRowSizeInBytes*1024L*1024L; // 1 million rows
     } else {
       Long size = sizeMap.get(regionId);
       if (size == null) {
